Linting + Restructure to fix circular import
