+++ lenny/embeds/sessionstats.py
@@ -1,9 +1,8 @@
 from embeds.embed import BaseEmbed
-from logic.sessionsstats import UserSessionResult
+from logic.session.stats import UserSessionResult
 
 
 class UserSessionStatEmbed(BaseEmbed):
-
     def __init__(self, user_result: UserSessionResult):
         super().__init__(title=user_result.title, description=user_result.description, color=user_result.color)
